Bug fixed in edit form for title and reorganized the notes crud operations
